new results (total 200) with minor code update

+++ aRunSim.R
@@ -57,12 +57,8 @@ save_eval <- function(benchmarks, id){
 
   Sys.sleep(runif(1))
   fname <- paste0(path, "benchmarks-", id, ".rds")
-  if (file.exists(fname)) {
-    saveRDS(rbind(readRDS(fname), df), file = fname)
-    
-  } else {
-    saveRDS(df, file = fname)
-  }
+  if (file.exists(fname)) saveRDS(rbind(readRDS(fname), df), file = fname)
+  else  saveRDS(df, file = fname)
 }
 
 
@@ -95,20 +91,30 @@ save_eval <- function(benchmarks, id){
 # boosting ----------------------------------------------------------------
 
 
-for(i in 1:100){
-  Omega <- matrix(c(2, -9, -4, -6, 5, 6, 0.5, 3, 8), ncol =3)
-  Epsilon <- 1e-5 * matrix(c(-3, -4, 9, 6, 2.5, 2.1, -6, 6, 7), ncol = 3)
-  zall <- sim_good_sources(N = 1e4, 3)
-  xall <- tvmix(zall, Omega, Epsilon)
-  
-  # loop for freqs
-  freq_list <- 2^(0:10)
-  for(freq in freq_list){
-    for(l in c(3,6,12,1)){
-      ids  <- seq(from = 1, to = nrow(xall), by = freq)
-      x <- xall[ids,]
-      z <- zall[ids,]
-      do_it_once(x, z, lll = l, id = paste0("fixed_freq_E5N4_Boot_lag", l))
+multido <- function(E, N, sn){
+  for(i in 1:50){
+    Omega <- matrix(c(2, -9, -4, -6, 5, 6, 0.5, 3, 8), ncol =3)
+    Epsilon <- 10^(-E) * matrix(c(-3, -4, 9, 6, 2.5, 2.1, -6, 6, 7), ncol = 3)
+    zall <- sim_good_sources(N = 1e4, 3)
+    xall <- tvmix(zall, Omega, Epsilon)
+    
+    # loop for freqs
+    freq_list <- 2^(0:10)
+    for(freq in freq_list){
+      for(l in c(3,6,12,1)){
+        ids  <- seq(from = 1, to = nrow(xall), by = freq)
+        x <- xall[ids,]
+        z <- zall[ids,]
+        do_it_once(x, z, lll = l, id = paste0("seq", sn, "_fixed_freq_E", E, "N", N, "_Boot_lag", l))
+      }
     }
   }
 }
+
+
+
+library(parallel)
+mclapply(as.list(1:100), function(seq) {
+  multido(5,5,seq); multido(5,4,seq);
+  multido(4,5,seq); multido(4,4,seq);
+})
